Update address in geyserconnect.tag
Updates the address from `geyserconnect.gq` to`geyserconnect.net`

From GeyserConnect discord: 

"Big news guys! We are migrating from the current domain name to `geyserconnect.net`! We haven't transitioned fully yet to the new domain, but it is usable if you wish to give it a try! We're aiming to have it fully migrated sometime next month."