Make logs.tag more forward
Camotoy also suggested this: 

```
An entire log file is a very useful way    of debugging any issues    with plugins. Sometimes    a snippet of a log doesn't tell    the whole story!

You can upload your log file into this channel, or you can use https://mclo.gs/ to censor all IPs present in your log. 
For BungeeCord (not Waterfall) there is a `proxy.log.0` file in the root and for everything else it is `latest.log` in `logs` folder.
```